requirements.txt
kafka-python==1.4.7
logstash-formatter==0.5.8
